fix: review component화
